Refactor for extensibility
